feat: emoji display picture can be changed
